:truck: Move S3 modules to FileManager
