+++ src/FileManager/aws-sdk.config.ts
@@ -4,7 +4,7 @@ env.config();
 export const AWS_SDK_CONFIG = {
   credentials: {
     accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-    secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY
+    secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
   },
-  region: process.env.AWS_REGION
+  region: process.env.AWS_REGION,
 };
