+++ src/drivers/drivers.ts
@@ -4,7 +4,7 @@ import { ExpressAuthRouteDriver } from './express/ExpressAuthRouteDriver';
 import { ExpressAdminRouteDriver } from './express/ExpressAdminRouteDriver';
 import { ExpressResponder } from './express/ExpressResponder';
 import { MongoDriver } from './MongoDriver';
-import { S3Driver } from './S3Driver';
+import { S3Driver } from '../FileManager/S3Driver';
 export {
   ExpressDriver,
   ExpressRouteDriver,
